+++ packages/kit/test/apps/basics/src/routes/remote/form/form.remote.ts
@@ -14,7 +14,8 @@ export const set_message = form(
 		message: v.picklist(
 			['hello', 'goodbye', 'unexpected error', 'expected error', 'redirect'],
 			'message is invalid'
-		)
+		),
+		uppercase: v.optional(v.string())
 	}),
 	async (data) => {
 		if (data.message === 'unexpected error') {
@@ -29,7 +30,7 @@ export const set_message = form(
 			redirect(303, '/remote');
 		}
 
-		message = data.message;
+		message = data.uppercase === 'true' ? data.message.toUpperCase() : data.message;
 
 		if (getRequestEvent().isRemoteRequest) {
 			const deferred = Promise.withResolvers();