+++ packages/kit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @sveltejs/kit
 
+## 2.43.3
+### Patch Changes
+
+
+- fix: Webcontainer AsyncLocalStorage workaround ([#14521](https://github.com/sveltejs/kit/pull/14521))
+
+
+- fix: include the value of form submitters on `form` remote functions ([#14475](https://github.com/sveltejs/kit/pull/14475))
+
 ## 2.43.2
 ### Patch Changes